+++ CHANGELOG.md
@@ -25,6 +25,13 @@ Add a link to the GitHub diff like
 [Dev Changes](https://github.com/mastersign/bench/compare/master...dev),
 [App Changes](https://github.com/mastersign/bench/compare/master...apps)
 
+### Changed
+- Changed app ID of 7-Zip from `SvZ` to `7z`
+
+### Fixed
+- Typo in `HTTP_PROXY` and `HTTPS_PROXY` in `env.cmd`
+  ([#71](https://github.com/mastersign/bench/issues/71))
+
 ## [0.11.4] - 2016-07-04
 [Full Changelog](https://github.com/mastersign/bench/compare/v0.11.3...v0.11.4)