+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Add a link to the GitHub diff like
 ### Changed
 - Changed app ID of 7-Zip from `SvZ` to `7z`
 - Update: JDK/JRE 8 from rev. 92 to rev. 102
+- Update: JabRef from 3.3 to 3.5
 
 ### Fixed
 - Typo in `HTTP_PROXY` and `HTTPS_PROXY` in `env.cmd`